pyproject.toml
[project]
name = "calisthenics-coach"
version = "1.0.0"
description = "Real-time AI coach for bodyweight exercises using YOLO pose detection and Gemini voice feedback"
authors = [
    {name = "Kartik Pal", email = "palkartik0307@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ai", "computer-vision", "fitness", "yolo", "pose-detection", "gemini", "vision-agents"]

dependencies = [
    "vision-agents[getstream,gemini,ultralytics]>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/palkartik07/calisthenics-coach"
Repository = "https://github.com/palkartik07/calisthenics-coach"
"Bug Reports" = "https://github.com/palkartik07/calisthenics-coach/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["."]